Expose GML3 writing as GML3Writer.
